test: Add setup script for /test/tmp
I'll have one subfolder per config, and copy the files from the fixture in the tmp folder, so that it onyl has to install once per config after all.

It's not as clean as installing the dependencies in each fixture, but it's an acceptable compromise to make tests run faster as it will only have to `npm install` three times (once toplevel, once in each of the two tmp folders).